refactor: 로컬 최신화 후 중복 코드 제거 & import 경로 수정
