--- src/main/java/com/commerceteamproject/admin/enitity/AdminRole.java
@@ -1,7 +1,0 @@
-package com.commerceteamproject.admin.enitity;
-
-public enum AdminRole {
-    // 슈퍼 관리자, 운영 관리자, CS 관리자
-    SUPER, RUN, CS
-}
-
