--- src/main/java/com/commerceteamproject/admin/enitity/AdminStatus.java
@@ -1,6 +1,0 @@
-package com.commerceteamproject.admin.enitity;
-
-public enum AdminStatus {
-    // 활성화, 비활성화, 정지, 승인대기, 거부
-    ACTIVATION, INACTIVATION, SUSPENSION, PENDING, REJECTION
-}
