--- src/main/java/com/commerceteamproject/admin/관리자 정보 관리.txt
@@ -1,53 +1,0 @@
-관리자 정보 관리
-->  슈퍼 관리자가 등록된 관리자 계정을 조회, 수정, 삭제하고, 신규 관리자의 가입 신청을 승인 또는 거부할 수 있도록 합니다.
-< 관리자 리스트 조회 >
-쿼리 파라미터 지원
-	- 검색 키워드(이름, 이메일)
-	- 페이지 번호(기본값 : 1)
-	- 페이지당 개수(기본값 10)
-	- 정렬 기준(이름)
-	- 정렬 순서(오름차순)
-	- 역할 필터 : 슈퍼관리자, 운영관리자, CS관리자 ( 일단은 슈퍼만 )
-	- 상태 필터 : 활성/비활성, 정지, 승인대기, 거부
-    응답 포함
-        - 관리자 목록(고유 식별자 ID, 이름, 이메일, 전화번호, 역할, 상태, 가입일/승인일(BaseEntity)
-        - 페이징 정보(현재 페이지, 페이지당 개수, 전체 개수, 전체 페이지 수)
-
-✅ [관리자 상세 조회]
-    - 특정 관리자의 상세 정보를 조회합니다.
-        - `이름`, `이메일`, `전화번호`, `역할`, `상태`, `가입일`, `승인일`
-    - 존재하지 않는 ID 요청 시 에러를 반환합니다.
-
-✅ [관리자 정보 수정]
-    - 수정 가능한 필드: `이름`, `이메일`, `전화번호`
-
-[관리자 역할 변경]
-    - 관리자의 `역할`을 변경합니다.
-
-[관리자 상태 변경]
-    - 관리자의 `상태`를 변경합니다.
-
-✅ [관리자 삭제]
-    - 특정 관리자를 탈퇴(삭제)시킵니다.
-
-[관리자 승인/거부]
-    관리자 승인:
-        - `승인대기` 상태의 관리자를 `활성` 상태로 변경합니다.
-        - 승인 일시 필드를 업데이트합니다.
-
-    관리자 거부:
-        - `승인대기` 상태의 관리자를 `거부` 상태로 변경합니다.
-        - 거부 일시, 거부 사유 필드를 업데이트합니다.
-        - 거부 사유는 필수 입력 항목입니다.
-
-✅ [비밀번호 변경]
-    - 현재 비밀번호를 새로운 비밀번호로 변경합니다.
-
-------- 곽현민 담당 --------
-[내 프로필 조회]
-- 내(로그인한 유저) 프로필 정보를 조회할 수 있습니다.
-    - `이름`, `이메일`, `전화번호`
-
-[내 프로필 수정]
-- 수정 가능한 필드: `이름`, `이메일`, `전화번호`
----------------------------
