--- src/main/java/com/commerceteamproject/common/BaseEntity.java
@@ -1,23 +1,0 @@
-package com.commerceteamproject.common;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-
-public abstract class BaseEntity {
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-
-}
