--- src/main/java/com/commerceteamproject/common/ServiceException.java
@@ -1,13 +1,0 @@
-package com.commerceteamproject.common;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class ServiceException extends RuntimeException {
-    private final HttpStatus status;
-    public ServiceException(String message, HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-}
